requirements.txt
django>=2.0.4,<3.0.0
pyjwt==1.7.1
psycopg2-binary==2.8.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0